Update to instructions and added case normalization post processing.

+++ include/ollama.h
@@ -46,7 +46,7 @@ namespace sc {
         void setKeepAlive(std::string keep_alive) { this->keep_alive = keep_alive; }
         void setMaxTokens(int max_tokens) { this->max_tokens = max_tokens; }
         void setTopP(double top_p) { this->top_p = top_p; }
-        void setTopK(double top_k) { this->top_k = top_k; }
+        void setTopK(int top_k) { this->top_k = top_k; }
         void setRepeatPenalty(double repeat_penalty) { this->repeat_penalty = repeat_penalty; }
         void setNumPredict(double num_predict) { this->num_predict = num_predict; }
         void setNumCtx(double num_ctx) { this->num_ctx = num_ctx; }
@@ -66,7 +66,7 @@ namespace sc {
         std::string keep_alive{"30m"};
         int max_tokens{DEFAULT_MAX_TOKENS};
         float top_p{1.0};
-        float top_k{1.0};
+        int top_k{1};
         float repeat_penalty{1.0};
         int num_predict{4096};
         int num_ctx{8192};
